requirements.txt
Flask==3.1.2
Flask-WTF==1.2.2
supabase==2.28.0
firebase_admin==7.1.0
python-dotenv==1.2.2
psycopg2-binary==2.9.11
requests==2.32.5
WTForms==3.2.1
Werkzeug==3.1.5
Jinja2==3.1.6
cryptography==46.0.4
PyJWT==2.10.1
markdown-it-py==4.0.0
rich==14.3.3
